update items fix deletion

+++ public/js/controllers/admin-controller.js
@@ -2,35 +2,39 @@
 
 /** Controllers */
 angular.module('store.controllers')
-    .controller('AdminController', function ($scope, $modal, $route, storeItems) {
+    .controller('AdminController', function ($scope, $modal, $route, LoginService, storeItems) {
 
-        storeItems.getAllItemsByType('', function (data) {
-            $scope.items = data;
-        });
+        if (LoginService.isLoggedIn()) {
 
-        // for add item modal
-        $scope.open = function () {
-            var modalInstance = $modal.open({
-                templateUrl: 'public/partials/dialogs/add-item-dialog.html',
-                controller: 'AddItemController',
-                resolve: {
-                    items: function () {
-                        return null;
+            storeItems.getAllItemsByType('', function (data) {
+                $scope.items = data;
+            });
+
+            // for add item modal
+            $scope.open = function (itemId) {
+                var modalInstance = $modal.open({
+                    templateUrl: 'public/partials/dialogs/add-item-dialog.html',
+                    controller: 'AddItemController',
+                    resolve: {
+                        selected: function () {
+                            return itemId;
+                        }
                     }
-                }
-            });
-            modalInstance.result.then(function (selectedItem) {
-                $scope.selected = selectedItem;
-            }, function () {
-                console.log('Modal dismissed at: ' + new Date());
-            });
-        };
+                });
+//                modalInstance.result.then(function (selectedItem) {
+//                    $scope.selected = selectedItem;
+//                }, function () {
+//                    console.log('Modal dismissed at: ' + new Date());
+//                });
+            };
 
-        $scope.deleteItem = function (itemId) {
-            storeItems.deleteItem(itemId, function () {
-                $route.reload();
-                alert("Удаление прошло успешно.")
-            });
+            $scope.deleteItem = function (itemId) {
+                storeItems.deleteItem(itemId, function () {
+                    $route.reload();
+                    alert("Удаление прошло успешно.")
+                });
+            }
+
         }
 
     });
